refactor: extract new file `default-backend.ts`
To store default backend specific functions.
